Aliasing and command verification
Rewrote the command verification process to now test hashes against a dictionary so renamed files should still work. Aliasing is now supported, quotes are able to be used if the aliased command is more than 1 word long. Changed sysCont.userSystem to point to the user's system object instead of just the IP. Completely removed the connect and disconnect commands as well as the sysCont system switch routine.